allows you to pick any player if you'd only be allowed to pick insider but you aren't an insider
